pkg/providers/discoverers/cloud_foundry/test_data/test-app/requirements.txt
flask==2.3.3
mysql-connector-python